+++ minos/common/model/__init__.py
@@ -1,34 +1,18 @@
 from .abc import (
     Model,
 )
-from .actions import (
-    Action,
-)
 from .declarative import (
-    Aggregate,
-    AggregateDiff,
-    AggregateRef,
     Command,
     CommandReply,
     CommandStatus,
     DeclarativeModel,
-    Entity,
-    EntitySet,
     Event,
-    IncrementalSet,
-    IncrementalSetDiff,
-    IncrementalSetDiffEntry,
     MinosModel,
-    ValueObject,
-    ValueObjectSet,
 )
 from .dynamic import (
     BucketModel,
     DataTransferObject,
     DynamicModel,
-    FieldDiff,
-    FieldDiffContainer,
-    IncrementalFieldDiff,
 )
 from .fields import (
     Field,
@@ -43,11 +27,9 @@ from .serializers import (
 from .types import (
     GenericTypeProjector,
     MissingSentinel,
-    ModelRef,
-    ModelRefExtractor,
-    ModelRefInjector,
     ModelType,
     NoneType,
     TypeHintBuilder,
     TypeHintComparator,
+    is_model_type,
 )