+++ minos/common/model/abc.py
@@ -57,9 +57,12 @@ logger = logging.getLogger(__name__)
 class Model(Mapping):
     """Base class for ``minos`` model entities."""
 
-    _fields: dict[str, Field]
+    _field_cls: Type[Field] = Field
 
-    def __init__(self, fields: Union[Iterable[Field], dict[str, Field]] = None):
+    _fields: dict[str, Field]
+    __eq_reversing: bool
+
+    def __init__(self, fields: Union[Iterable[Field], dict[str, Field]] = None, **kwargs):
         """Class constructor.
 
         :param fields: Dictionary that contains the ``Field`` instances of the model indexed by name.
@@ -69,6 +72,7 @@ class Model(Mapping):
         if not isinstance(fields, dict):
             fields = {field.name: field for field in fields}
         self._fields = fields
+        self.__eq_reversing = False
 
     @classmethod
     def from_avro_str(cls: Type[T], raw: str, **kwargs) -> Union[T, list[T]]:
@@ -220,14 +224,14 @@ class Model(Mapping):
 
     def __setattr__(self, key: str, value: Any) -> None:
         if key.startswith("_"):
-            super().__setattr__(key, value)
+            object.__setattr__(self, key, value)
         elif key in self._fields:
             self._fields[key].value = value
         else:
             raise AttributeError(f"{type(self).__name__!r} does not contain the {key!r} field")
 
     def __getattr__(self, item: str) -> Any:
-        if item in self._fields:
+        if item != "_fields" and item in self._fields:
             return self._fields[item].value
         else:
             raise AttributeError(f"{type(self).__name__!r} does not contain the {item!r} field.")
@@ -269,7 +273,17 @@ class Model(Mapping):
         return MinosAvroProtocol().encode(self.avro_data, self.avro_schema)
 
     def __eq__(self: T, other: T) -> bool:
-        return type(self) == type(other) and self.fields == other.fields
+        if type(self) == type(other) and self.fields == other.fields:
+            return True
+
+        if not self.__eq_reversing:
+            try:
+                self.__eq_reversing = True
+                return other == self
+            finally:
+                self.__eq_reversing = False
+
+        return False
 
     def __hash__(self) -> int:
         return hash(tuple(self.fields.values()))