+++ minos/common/model/declarative/aggregate/__init__.py
@@ -1,9 +1,0 @@
-from .diff import (
-    AggregateDiff,
-)
-from .model import (
-    Aggregate,
-)
-from .ref import (
-    AggregateRef,
-)