+++ minos/common/model/dynamic/dto.py
@@ -49,8 +49,4 @@ class DataTransferObject(DynamicModel):
         return f"{self._name}[DTO]({fields_repr})"
 
     def __eq__(self, other: DataTransferObject):
-        return (
-            super(DynamicModel, self).__eq__(other)
-            and self._name == other._name
-            and self._namespace == other._namespace
-        )
+        return super().__eq__(other) and self._name == other._name and self._namespace == other._namespace