+++ minos/common/model/fields.py
@@ -141,7 +141,7 @@ class Field:
         return AvroSchemaEncoder(self.type, self.name).build()
 
     @property
-    def avro_data(self):
+    def avro_data(self) -> Any:
         """Compute the avro data of the model.
 
         :return: A dictionary object.
@@ -177,7 +177,7 @@ class Field:
         yield from (self.name, self.type, self.value, self._parser_function, self._validator_function)
 
     def __repr__(self) -> str:
-        return f"{self.name}={self.value!s}"
+        return f"{self.name}={self.value!r}"
 
 
 ModelField = Field